pin ansible and molecule for now
we have our tests failing thanks to a molecule and testinfra
incompatibility problem
